requirements.txt
annotated-doc==0.0.4 ; python_version >= "3.11"
annotated-types==0.7.0 ; python_version >= "3.11"
anyio==4.13.0 ; python_version >= "3.11"
attrs==26.1.0 ; python_version >= "3.11"
certifi==2026.2.25 ; python_version >= "3.11"
cffi==2.0.0 ; python_version >= "3.11" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.7 ; python_version >= "3.11"
click==8.3.2 ; python_version >= "3.11"
colorama==0.4.6 ; (platform_system == "Windows" or sys_platform == "win32") and python_version >= "3.11"
cryptography==46.0.6 ; python_version >= "3.11"
distro==1.9.0 ; python_version >= "3.11"
dnspython==2.8.0 ; python_version >= "3.11"
email-validator==2.3.0 ; python_version >= "3.11"
faker==40.12.0 ; python_version >= "3.11"
fastapi-cli==0.0.24 ; python_version >= "3.11"
fastapi-cloud-cli==0.15.1 ; python_version >= "3.11"
fastapi-mcp==0.4.0 ; python_version >= "3.11"
fastapi==0.135.3 ; python_version >= "3.11"
fastar==0.9.0 ; python_version >= "3.11"
google-auth==2.49.1 ; python_version >= "3.11"
google-genai==1.70.0 ; python_version >= "3.11"
greenlet==3.3.2 ; python_version >= "3.11" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.16.0 ; python_version >= "3.11"
httpcore==1.0.9 ; python_version >= "3.11"
httptools==0.7.1 ; python_version >= "3.11"
httpx-sse==0.4.3 ; python_version >= "3.11"
httpx==0.28.1 ; python_version >= "3.11"
idna==3.11 ; python_version >= "3.11"
jinja2==3.1.6 ; python_version >= "3.11"
jsonschema-specifications==2025.9.1 ; python_version >= "3.11"
jsonschema==4.26.0 ; python_version >= "3.11"
markdown-it-py==4.0.0 ; python_version >= "3.11"
markupsafe==3.0.3 ; python_version >= "3.11"
mcp==1.27.0 ; python_version >= "3.11"
mdurl==0.1.2 ; python_version >= "3.11"
numpy==2.4.4 ; python_version >= "3.11"
pandas==3.0.2 ; python_version >= "3.11"
pyasn1-modules==0.4.2 ; python_version >= "3.11"
pyasn1==0.6.3 ; python_version >= "3.11"
pycparser==3.0 ; platform_python_implementation != "PyPy" and implementation_name != "PyPy" and python_version >= "3.11"
pydantic-core==2.41.5 ; python_version >= "3.11"
pydantic-extra-types==2.11.1 ; python_version >= "3.11"
pydantic-settings==2.13.1 ; python_version >= "3.11"
pydantic==2.12.5 ; python_version >= "3.11"
pygments==2.20.0 ; python_version >= "3.11"
pyjwt==2.12.1 ; python_version >= "3.11"
python-dateutil==2.9.0.post0 ; python_version >= "3.11"
python-dotenv==1.2.2 ; python_version >= "3.11"
python-multipart==0.0.22 ; python_version >= "3.11"
pywin32==311 ; python_version >= "3.11" and sys_platform == "win32"
pyyaml==6.0 ; python_version >= "3.11"
referencing==0.37.0 ; python_version >= "3.11"
requests==2.33.1 ; python_version >= "3.11"
rich-toolkit==0.19.7 ; python_version >= "3.11"
rich==14.3.3 ; python_version >= "3.11"
rignore==0.7.6 ; python_version >= "3.11"
rpds-py==0.30.0 ; python_version >= "3.11"
sentry-sdk==2.57.0 ; python_version >= "3.11"
shellingham==1.5.4 ; python_version >= "3.11"
six==1.17.0 ; python_version >= "3.11"
sniffio==1.3.1 ; python_version >= "3.11"
sqlalchemy==2.0.49 ; python_version >= "3.11"
sse-starlette==3.3.4 ; python_version >= "3.11"
starlette==1.0.0 ; python_version >= "3.11"
tenacity==9.1.4 ; python_version >= "3.11"
tomli==2.4.1 ; python_version >= "3.11"
typer==0.24.1 ; python_version >= "3.11"
typing-extensions==4.15.0 ; python_version >= "3.11"
typing-inspection==0.4.2 ; python_version >= "3.11"
tzdata==2026.1 ; (sys_platform == "win32" or sys_platform == "emscripten" or platform_system == "Windows") and python_version >= "3.11"
urllib3==2.6.3 ; python_version >= "3.11"
uvicorn==0.42.0 ; python_version >= "3.11"
uvloop==0.22.1 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.11"
watchfiles==1.1.1 ; python_version >= "3.11"
websockets==16.0 ; python_version >= "3.11"
